Replace existing prod links with new flow
