+++ packages/common/src/hooks/index.ts
@@ -32,6 +32,7 @@ export * from './useChangePasswordFormConfiguration'
 export * from './useChangeEmailFormConfiguration'
 export * from './useCanSendChatBlast'
 export * from './useAudioRewards'
+export * from './useDiscordOAuthLink'
 
 export * from './useAccountSwitcher'
 export * from './useAccountHasClaimableRewards'
