+++ packages/common/src/store/ui/modals/types.ts
@@ -73,7 +73,6 @@ export type Modals =
   | 'BuyAudio'
   | 'BuyAudioRecovery'
   | 'TransactionDetails'
-  | 'VipDiscord'
   | 'StripeOnRamp'
   | 'CoinflowOnramp'
   | 'InboxSettings'
