--- packages/mobile/src/components/vip-discord-drawer/index.ts
@@ -1,1 +1,0 @@
-export * from './VipDiscordDrawer'
